fix(core/tag): fix type mismatch in BaseFakeTag.getHandle
- Convert getHandle return value from String/Long to Int (lastHandle.toInt())
- Ensure consistency with getTechHandles IntArray usage

ci:
- Add `id: attest` to Attest Build Provenance step
- Copy attestation bundle to release_assets/attestation.jsonl
- Remove checksums_sha256.txt during release asset cleanup

+++ app/src/main/java/mba/vm/onhit/core/tag/BaseFakeTag.kt
@@ -69,7 +69,7 @@ abstract class BaseFakeTag {
                     "getTechList" -> techList
                     "getTechExtras" -> techExtras
                     "isPresent" -> true
-                    "getHandle" -> lastHandle
+                    "getHandle" -> lastHandle.toInt()
                     "getTechHandles" -> IntArray(techList.size) { lastHandle.toInt() }
                     else -> {
                         when (method.returnType) {
